richtext insert, color renderer
